optimize performance

+++ include/rnnoise.h
@@ -41,7 +41,7 @@
 
 typedef struct DenoiseState DenoiseState;
 
-RNNOISE_EXPORT int rnnoise_get_size();
+RNNOISE_EXPORT size_t rnnoise_get_size();
 
 RNNOISE_EXPORT int rnnoise_init(DenoiseState *st);
 
@@ -49,4 +49,4 @@ RNNOISE_EXPORT DenoiseState *rnnoise_create();
 
 RNNOISE_EXPORT void rnnoise_destroy(DenoiseState *st);
 
-RNNOISE_EXPORT float rnnoise_process_frame(DenoiseState *st, float *out, const float *in);
+RNNOISE_EXPORT float rnnoise_process_frame(DenoiseState *st, short *out, const short *in);
